refactor: create hook for adding new pipeline steps
This functionality is needed in a different component now.
