+++ services/orchest-webserver/client/src/pipeline-view/contexts/PipelineEditorContext.tsx
@@ -32,7 +32,8 @@ export type PipelineEditorContextType = {
   eventVars: EventVars;
   dispatch: (value: EventVarsAction) => void;
   stepDomRefs: React.MutableRefObject<Record<string, HTMLDivElement | null>>;
-  pipelineCanvasRef: React.MutableRefObject<HTMLDivElement | null>;
+  pipelineCanvasRef: React.MutableRefObject<HTMLElement | null>;
+  pipelineViewportRef: React.RefObject<HTMLElement | null>;
   newConnection: React.MutableRefObject<NewConnection | undefined>;
   keysDown: Set<number | string>;
   trackMouseMovement: (clientX: number, clientY: number) => void;
@@ -80,13 +81,13 @@ export const PipelineEditorContextProvider: React.FC = ({ children }) => {
     state: { pipelines, projectUuid, pipeline },
   } = useProjectsContext();
   const pipelineUuid = pipeline?.uuid;
-
   const { jobUuid, runUuid: runUuidFromRoute } = useCustomRoute();
 
   // No pipeline found. Editor is frozen and shows "Pipeline not found".
   const disabled = hasValue(pipelines) && pipelines.length === 0;
 
   const pipelineCanvasRef = React.useRef<HTMLDivElement | null>(null);
+  const pipelineViewportRef = React.useRef<HTMLDivElement | null>(null);
 
   const {
     eventVars,
@@ -199,6 +200,7 @@ export const PipelineEditorContextProvider: React.FC = ({ children }) => {
         dispatch,
         stepDomRefs,
         pipelineCanvasRef,
+        pipelineViewportRef,
         newConnection,
         keysDown,
         trackMouseMovement,
